Provide a unique copy of options to each call to compile()
The compiler adds attributes to the options object on the first pass;
consequential passes are not decorated with the '__extend()' and other
functions.  This makes sense when compiling a long list of javascript
files into a single file, but grunt-coffee compiles each file into
its javascript equivalent at the target end.

This patch gives a fresh clone of the options object passed from the
user's grunt.js configuration file to each call to coffee.compile().